+++ pmd-core/src/main/java/net/sourceforge/pmd/RuleSetLoader.java
@@ -25,7 +25,7 @@ import net.sourceforge.pmd.annotation.InternalApi;
 import net.sourceforge.pmd.lang.Language;
 import net.sourceforge.pmd.lang.LanguageRegistry;
 import net.sourceforge.pmd.util.CollectionUtil;
-import net.sourceforge.pmd.util.ResourceLoader;
+import net.sourceforge.pmd.util.internal.ResourceLoader;
 import net.sourceforge.pmd.util.log.PmdReporter;
 
 /**